Move the warnings/errors in a place that doesn't import anything.
